product-service-2/requirements.txt
fastapi
uvicorn
sqlalchemy
pydantic
psycopg2-binary
PyJWT